adding support for template refs:
As part of the list of templates, this PR is adding support to have an entry like:

{
    "ref": "https://url/templates.json",
    "tags": ["aiApps"]
}

When this entry is found, awesome-azd will expand that entry into all the entries referenced on that url. Then, the tags are appended to the tags from the original entries.

This feature allow a seemly experience to batch a group of templates which are already maintained as part of another source and avoid duplicity.