add `nlp-sql-in-a-box` template (#419)
* feat: Add Azure Speech Services tag

The commit adds a new tag for Azure Speech Services in the `tags.tsx` file. This tag includes the label, description, Azure icon, URL, and type for the service.

* feat: Add NLP to SQL template and update other projects with 'speechservice' tag.

The commit adds the NLP to SQL template to the `templates.json` file. This template is part of the AI-in-a-Box framework and provide ready-to-use accelerators for AI and ML solutions.
Also added 'speechservice' to other projects using it.